requirements.txt
# Core ML and AI libraries
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
accelerate>=0.12.0
huggingface_hub>=0.8.0
timm>=0.6.0

# Computer Vision
opencv-python>=4.6.0
Pillow>=9.0.0

# Audio Processing
librosa>=0.9.0
pyaudio>=0.2.11
sounddevice>=0.4.5
soundfile>=0.10.0
pygame>=2.1.0

# Data Processing
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
scipy>=1.8.0

# GUI and Interface
tk>=0.1.0

# Utilities
requests>=2.28.0
joblib>=1.1.0
matplotlib>=3.5.0

# Document Processing
PyPDF2>=2.0.0
python-docx>=0.8.11

# Windows Specific (if needed)
pywin32>=303; sys_platform == "win32"

# Optional: For better performance
# tensorrt>=8.4.0  # NVIDIA GPU acceleration
# onnxruntime>=1.12.0  # ONNX runtime for inference